more Material Design 3 stuff (and extracted composables to external functions (other files))
