+++ app/src/main/java/de/schweininchen/shintaikan/shintaikan/jetpack/pages/FF15.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -49,7 +48,7 @@ fun FirebaseDataPage(
                 // Enable the scale animation
                 scale = true,
                 // Change the color and shape
-                backgroundColor = MaterialTheme.colors.primary,
+                backgroundColor = androidx.compose.material3.MaterialTheme.colorScheme.primary,
             )
         }
